% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_Codons.R
\name{mk_codons}
\alias{mk_codons}
\title{R/mk_codons.R
Making a Codonframe
Takes a DNA Sequence and returns a list of triplets (Codons)}
\usage{
mk_codons(dna, s = 1)
}
\arguments{
\item{s}{start of making codons in sequence}

\item{DNA}{sequence}
}
\value{
codons
}
\description{
R/mk_codons.R
Making a Codonframe
Takes a DNA Sequence and returns a list of triplets (Codons)
}
