% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_codons_to_aa.R
\name{dna_codons_to_aa}
\alias{dna_codons_to_aa}
\title{DNA_to_Codons}
\usage{
dna_codons_to_aa(codons)
}
\arguments{
\item{Codons}{you want translated.}
}
\value{
Amino Acids
}
\description{
Translates codons to the correct amino acid.
}
