% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_Codons.R
\name{mk_codons}
\alias{mk_codons}
\title{Making a Codonframe
Takes a DNA Sequence and returns a list of triplets (Codons)}
\usage{
mk_codons(dna, s = 1)
}
\arguments{
\item{s}{start of making codons in sequence
return codons}

\item{DNA}{sequence}
}
\description{
Making a Codonframe
Takes a DNA Sequence and returns a list of triplets (Codons)
}
