% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_dna.R
\name{random_dna}
\alias{random_dna}
\title{Takes a length and returns a DNA string of equal length}
\usage{
random_dna(l)
}
\arguments{
\item{length}{of wanted/expected DNA string}
}
\description{
Takes a length and returns a DNA string of equal length
}
